Move aggregation to module
